requirements.txt
flask
graphviz
python-dotenv
google-genai
markdown